requirements.txt
langchain-astradb
langchain 
langchain-openai
datasets
pypdf
python-dotenv
flask
google-generativeai
langchain-google-genai


-e .